+++ src/utils.ts
@@ -88,7 +88,7 @@ export async function sanitizeHtmlAllowingATag(dirtyHtml: string): Promise<strin
   });
 }
 
-export async function encrypt(bytes: Uint8Array, password: string): Promise<Uint8Array> {
+export async function encrypt(bytes: Uint8Array, password: string | Uint8Array): Promise<Uint8Array> {
   const openpgp = await openpgpAsync();
   // Encrypt with PGP
   const encryptResult = await openpgp.encrypt({
@@ -102,12 +102,28 @@ export async function encrypt(bytes: Uint8Array, password: string): Promise<Uint
   return encrypted;
 }
 
-export async function decrypt(bytes: Uint8Array, password: string): Promise<Uint8Array> {
+export async function decrypt(bytes: Uint8Array, password: string | Uint8Array): Promise<Uint8Array> {
   const openpgp = await openpgpAsync();
   const plain = (await openpgp.decrypt({
     message: await openpgp.message.read(bytes),
-    passwords: [password],
+    passwords: [password] as any, // TODO: Not use any
     format: 'binary'
   })).data as Uint8Array;
   return plain;
+}
+
+// (from: https://stackoverflow.com/a/40031979/2885946)
+export function buffToHex(buff: ArrayBuffer): string {
+  return Array.prototype.map.call(new Uint8Array(buff), (x) => ('00' + x.toString(16)).slice(-2)).join('');
+}
+
+export async function sha256(input: string): Promise<string> {
+  return buffToHex(
+    await crypto.subtle.digest('SHA-256', new TextEncoder().encode(input)),
+  );
+}
+
+// (from: https://stackoverflow.com/a/11562550/2885946)
+export function uint8ArrayToBase64(arr: Uint8Array): string {
+  return btoa(String.fromCharCode(...arr));
 }