+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.4.0] - 2019-11-19
+### Changed
+- Update dependencies
+- Merge data uploader and data viewer together
+
+## Added
+- Add passwordless protection 
+
 ## [0.3.1] - 2019-11-06
 ### Changed
 - Update dependencies
@@ -57,7 +65,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 - First release
 
-[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/nwtgck/piping-ui-web/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.1.5...v0.2.0